feat: auth login is function with private and public routes
