+++ src/pages/(public)/Login/_types/index.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export type LoginFormData = z.infer<typeof loginSchema>;
+
+export const loginSchema = z.object({
+	email: z.string().email('Digite um e-mail válido'),
+	password: z.string().min(3, 'A senha deve ter pelo menos 3 caracteres'),
+});
